Move tests into `ssz_derive`
